BalancePill to harmony-native
