+++ packages/mobile/src/harmony-native/components/index.ts
@@ -32,3 +32,4 @@ export * from './IconText'
 export * from './SendIcon'
 export * from './Artwork'
 export * from './Skeleton'
+export * from './BalancePill/BalancePill'
